+++ api/package.json
@@ -60,6 +60,7 @@
     "@types/express-session": "^1.17.4",
     "@types/google-libphonenumber": "^7.4.19",
     "@types/jsonwebtoken": "^8.5.5",
+    "@types/lodash": "^4.14.191",
     "@types/node": "16",
     "@types/node-mailjet": "^3.3.7",
     "@types/sinon": "^10.0.3",
