+++ api/proxy/package.json
@@ -32,6 +32,7 @@
     "@pdc/provider-sentry": "~0",
     "@pdc/provider-token": "~0",
     "@pdc/service-acquisition": "~0",
+    "@pdc/service-apdf": "~0",
     "@pdc/service-application": "~0",
     "@pdc/service-carpool": "~0",
     "@pdc/service-cee": "~0",
@@ -48,10 +49,10 @@
     "body-parser": "^1.19.0",
     "connect-redis": "^6.0.0",
     "cors": "^2.8.5",
-    "express": "^4.17.1",
     "express-mung": "^0.5.1",
     "express-rate-limit": "^6.7.0",
     "express-session": "^1.17.1",
+    "express": "^4.17.1",
     "helmet": "^4.4.1",
     "ioredis": "^4.27.9",
     "prom-client": "^14.0.0",
