+++ api/services/apdf/.gitignore
@@ -1,0 +1,3 @@
+dist
+coverage
+.nyc_output
