+++ api/services/apdf/ava.coverage.cjs
@@ -1,0 +1,5 @@
+const { coverage } = require('../../ava.common.cjs');
+
+module.exports = {
+    ...coverage,
+}
