+++ api/services/apdf/ava.integration.cjs
@@ -1,0 +1,5 @@
+const { integration } = require('../../ava.common.cjs');
+
+module.exports = {
+    ...integration,
+}
