+++ api/services/apdf/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "@pdc/service-trip",
+  "name": "@pdc/service-apdf",
   "version": "0.0.1",
   "private": true,
   "license": "Apache-2.0",
   "scripts": {
-    "ilos": "ilos",
     "build": "tsc",
+    "ilos": "ilos",
     "watch": "tsc -w",
     "test:unit": "ava --config ava.unit.cjs",
     "test:integration": "ava --config ava.integration.cjs",
@@ -23,18 +23,13 @@
     "@ilos/common": "~0",
     "@ilos/connection-manager": "~0",
     "@ilos/connection-postgres": "~0",
-    "@ilos/connection-redis": "~0",
     "@ilos/core": "~0",
+    "@ilos/framework": "~0",
     "@pdc/helper-test": "~0",
-    "@pdc/provider-file": "~0",
     "@pdc/provider-middleware": "~0",
     "@pdc/provider-validator": "~0",
-    "adm-zip": "^0.5.6",
-    "axios": "^1.2.2",
-    "csv-stringify": "^5.6.5",
-    "date-fns-tz": "^1.1.6",
-    "form-data": "^4.0.0",
-    "lodash": "^4.17.21",
-    "uuid": "^8.3.2"
+    "date-fns": "^2.29.3",
+    "date-fns-tz": "^1.3.7",
+    "exceljs": "^4.3.0"
   }
 }
