+++ api/services/apdf/src/ServiceProvider.integration.spec.ts
@@ -1,0 +1,5 @@
+import { serviceProviderMacro } from '@pdc/helper-test';
+import { ServiceProvider } from './ServiceProvider';
+
+const { test, boot } = serviceProviderMacro(ServiceProvider);
+test(boot);
