+++ api/services/apdf/src/bootstrap.ts
@@ -1,0 +1,7 @@
+import { Bootstrap as BaseBootstrap } from '@ilos/framework';
+
+import { ServiceProvider } from './ServiceProvider';
+
+export const bootstrap = BaseBootstrap.create({
+  serviceProviders: [ServiceProvider],
+});
