+++ api/services/apdf/src/config/ajv.ts
@@ -1,0 +1,4 @@
+export const config = {
+  allErrors: true,
+  verbose: true,
+};
