+++ api/services/apdf/src/config/index.ts
@@ -1,0 +1,7 @@
+import * as ajv from './ajv';
+import * as connections from './connections';
+
+export const config = {
+  ajv,
+  connections,
+};
