+++ api/services/apdf/src/interfaces/APDFTripInterface.ts
