+++ api/services/apdf/src/providers/excel/writer/SlicesWorkbookWriter.integration.spec.ts
@@ -3,8 +3,8 @@ import anyTest, { TestFn } from 'ava';
 import { stream, Workbook, Worksheet } from 'exceljs';
 import { SlicesWorkbookWriter } from './SlicesWorkbookWriter';
 import { BuildExcel } from '../BuildExcel';
-import { TripRepositoryProvider } from '../../../providers/TripRepositoryProvider';
-import { SliceStatInterface } from '../../../interfaces/PolicySliceStatInterface';
+import { SliceStatInterface } from '../../../shared/apdf/interfaces/PolicySliceStatInterface';
+import { APDFRepositoryProvider } from '../../../providers/APDFRepositoryProvider';
 
 interface Context {
   // Injected tokens
@@ -32,7 +32,7 @@ test('SlicesWorkbookWriter: should map slice into a dedicated worksheet', async 
   const slices: SliceStatInterface[] = [
     { slice: { start: 0, end: 2000 }, count: 2500, sum: 154588 },
     { slice: { start: 2000, end: 30000 }, count: 3000, sum: 204598 },
-    { slice: { start: 30000, end: TripRepositoryProvider.MAX_KM_LIMIT }, count: 5000, sum: 304456 },
+    { slice: { start: 30000, end: APDFRepositoryProvider.MAX_DISTANCE_METERS }, count: 5000, sum: 304456 },
   ];
 
   // Act
