+++ api/services/apdf/src/providers/excel/writer/SlicesWorkbookWriter.ts
@@ -1,9 +1,9 @@
 import { provider } from '@ilos/common';
 import { Column, stream, Worksheet } from 'exceljs';
-import { SliceStatInterface } from '../../../interfaces/PolicySliceStatInterface';
+import { APDFRepositoryProvider } from '../../../providers/APDFRepositoryProvider';
+import { SliceStatInterface } from '../../../shared/apdf/interfaces/PolicySliceStatInterface';
+import { SliceInterface } from '../../../shared/policy/common/interfaces/SliceInterface';
 import { AbstractWorkBookWriter } from './AbstractWorkbookWriter';
-import { TripRepositoryProvider } from '../../../providers/TripRepositoryProvider';
-import { SliceInterface } from '~/shared/policy/common/interfaces/SliceInterface';
 
 /***
  * Tranches       | Somme incentives  | Nombre trajet
@@ -37,7 +37,7 @@ export class SlicesWorkbookWriter extends AbstractWorkBookWriter {
   private formatSliceLabel(slice: SliceInterface): string {
     if (!slice.start && slice.end) {
       return `Jusqu'à ${slice.end / 1000} km`;
-    } else if (slice.end === TripRepositoryProvider.MAX_KM_LIMIT) {
+    } else if (slice.end === APDFRepositoryProvider.MAX_DISTANCE_METERS) {
       return `Supérieur à ${slice.start / 1000} km`;
     }
     return `De ${slice.start / 1000} km à ${slice.end / 1000} km`;
