+++ api/services/trip/src/actions/file/BuildFile.ts
@@ -6,7 +6,7 @@ import { normalizeExport, normalizeOpendata } from '../../helpers/normalizeExpor
 import fs from 'fs';
 import os from 'os';
 import { getOpenDataExportName } from '../../helpers/getOpenDataExportName';
-import { PgCursorHandler } from '../../interfaces/PromisifiedPgCursor';
+import { PgCursorHandler } from '../../shared/common/PromisifiedPgCursor';
 import { FormatInterface, ParamsInterface } from '../../shared/trip/buildExport.contract';
 import { BuildExportAction } from '../BuildExportAction';
 import { ExportTripInterface } from '~/interfaces';
