+++ api/services/trip/src/interfaces/TripRepositoryProviderInterface.ts
@@ -1,14 +1,11 @@
-import { SliceInterface } from '~/shared/policy/common/interfaces/SliceInterface';
 import { ResultWithPagination } from '../shared/common/interfaces/ResultWithPagination';
+import { PgCursorHandler } from '../shared/common/PromisifiedPgCursor';
 import { LightTripInterface } from '../shared/trip/common/interfaces/LightTripInterface';
 import {
   TripSearchInterface,
   TripSearchInterfaceWithPagination,
 } from '../shared/trip/common/interfaces/TripSearchInterface';
-import { APDFTripInterface } from './APDFTripInterface';
 import { ExportTripInterface } from './ExportTripInterface';
-import { PolicyStatsInterface } from './PolicySliceStatInterface';
-import { PgCursorHandler } from './PromisifiedPgCursor';
 import { FinancialStatInterface, StatInterface } from './StatInterface';
 
 export interface TzResultInterface {
@@ -39,9 +36,6 @@ export interface TripRepositoryInterface {
     type?: string,
   ): Promise<PgCursorHandler<ExportTripInterface>>;
   validateTz(tz?: string): Promise<TzResultInterface>;
-  getPolicyActiveOperators(campaign_id: number, start_date: Date, end_date: Date): Promise<number[]>;
-  getPolicyStats(params: CampaignSearchParamsInterface, slices: SliceInterface[]): Promise<PolicyStatsInterface>;
-  getPolicyCursor(params: CampaignSearchParamsInterface): Promise<PgCursorHandler<APDFTripInterface>>;
 }
 export abstract class TripRepositoryProviderInterfaceResolver implements TripRepositoryInterface {
   public async stats(params: Partial<TripSearchInterface>): Promise<StatInterface[]> {
@@ -80,21 +74,6 @@ export abstract class TripRepositoryProviderInterfaceResolver implements TripRep
   }
 
   public async getTerritoryDescendants(territory_id: number): Promise<number[]> {
-    throw new Error('Not implemented');
-  }
-
-  public async getPolicyActiveOperators(campaign_id: number, start_date: Date, end_date: Date): Promise<number[]> {
-    throw new Error('Not implemented');
-  }
-
-  public async getPolicyStats(
-    params: CampaignSearchParamsInterface,
-    slices: SliceInterface[],
-  ): Promise<PolicyStatsInterface> {
-    throw new Error('Not implemented');
-  }
-
-  public async getPolicyCursor(params: CampaignSearchParamsInterface): Promise<PgCursorHandler<APDFTripInterface>> {
     throw new Error('Not implemented');
   }
 }
