+++ api/services/trip/src/interfaces/index.ts
@@ -1,4 +1,3 @@
 export * from './TripRepositoryProviderInterface';
 export * from './ExportTripInterface';
-export * from './APDFTripInterface';
 export * from './DataGouvInterface';
